Moving Dataset and DatasetMetadata into own file
