+++ nexus_constructor/model/group.py
@@ -4,32 +4,9 @@ import attr
 
 from nexus_constructor.common_attrs import CommonAttrs
 from nexus_constructor.model.attribute import FieldAttribute
+from nexus_constructor.model.dataset import DatasetMetadata, Dataset
 from nexus_constructor.model.node import Node, _get_item, _set_item
 import numpy as np
-
-
-@attr.s
-class DatasetMetadata:
-    size = attr.ib(type=List)
-    type = attr.ib(type=str)
-
-
-@attr.s
-class Dataset(Node):
-    dataset = attr.ib(type=DatasetMetadata)
-    values = attr.ib(type=List, default=None)
-    type = attr.ib(type=str, default="dataset")
-    attributes = attr.ib(type=List[FieldAttribute], default=[])
-
-    @property
-    def nx_class(self):
-        return self.get_attribute_value(CommonAttrs.NX_CLASS)
-
-    @nx_class.setter
-    def nx_class(self, new_nx_class: str):
-        self.set_attribute_value(
-            CommonAttrs.NX_CLASS, new_nx_class,
-        )
 
 
 @attr.s
